+++ src/aioway/hop/__init__.py
@@ -1,3 +1,4 @@
 # Copyright (c) AIoWay Authors - All Rights Reserved
 
-from .inits import *
+from .hop import *
+from .nn import *
