+++ src/aioway/hop/hop.py
@@ -1,0 +1,4 @@
+# Copyright (c) AIoWay Authors - All Rights Reserved
+
+
+__all__ = []
