+++ src/aioway/hop/nn/embs.py
