+++ src/aioway/hop/nn/sliding.py
